Fix January display issue #98 and #99
Fixes  issue #98 and #99
display problem occurs for any year where January 1 is a Sunday and firstweekday is set to "sunday".  fixes code by checking if the current month is January and if January 1 of the current year falls on a Sunday. If so, it sets w to the week number returned by strftime instead of isocalendar and sets wn to zero. That fixes the January problem for years where January 1 is on a Sunday and the firstweekday = 'sunday' without messing up any other months. Code fix as proposed by @jemiele1 in #99