use 64bit python
Pillow 9.5.0 is the last Pillow version where they released 32-bit wheels. The solution to "Failed to build pillow" from missing zlib is to pin to old Pillow 9.5.0 or to use 64bit python. Opting for 64bit python.
This should fix the wheel build errors but won't fix the issues with the Babel parse error